+++ WalletWasabi.Gui/Tabs/WelcomeTabView/HardwareWallets/ConnectHardwareWalletView.xaml
@@ -3,7 +3,7 @@
              xmlns:controls="clr-namespace:WalletWasabi.Gui.Controls;assembly=WalletWasabi.Gui"
              xmlns:i="clr-namespace:Avalonia.Xaml.Interactivity;assembly=Avalonia.Xaml.Interactivity"
              xmlns:behaviors="clr-namespace:WalletWasabi.Gui.Behaviors;assembly=WalletWasabi.Gui"
-             x:Class="WalletWasabi.Gui.Tabs.WalletManager.HardwareWallets.ConnectHardwareWalletView">
+             x:Class="WalletWasabi.Gui.Tabs.WelcomeTab.HardwareWallets.ConnectHardwareWalletView">
   <DockPanel LastChildFill="True">
     <StackPanel DockPanel.Dock="Bottom" Spacing="10">
       <controls:GroupBox TextBlock.FontSize="30" Padding="30" Margin="20 0">
