+++ WalletWasabi.Gui/Tabs/WelcomeTabView/RecoverWallets/RecoverWalletView.xaml
@@ -4,7 +4,7 @@
              xmlns:i="clr-namespace:Avalonia.Xaml.Interactivity;assembly=Avalonia.Xaml.Interactivity"
              xmlns:behaviors="clr-namespace:WalletWasabi.Gui.Behaviors;assembly=WalletWasabi.Gui"
              xmlns:suggestions="clr-namespace:WalletWasabi.Gui.Suggestions;assembly=WalletWasabi.Gui"
-             x:Class="WalletWasabi.Gui.Tabs.WalletManager.RecoverWallets.RecoverWalletView">
+             x:Class="WalletWasabi.Gui.Tabs.WelcomeTab.RecoverWallets.RecoverWalletView">
   <StackPanel Margin="10" Spacing="10">
     <controls:GroupBox Title="{Binding Title}" TextBlock.FontSize="30" Padding="30">
       <Grid Classes="content">
