+++ WalletWasabi.Gui/Tabs/WelcomeTabView/WelcomeTabView.xaml
@@ -2,7 +2,7 @@
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:controls="clr-namespace:WalletWasabi.Gui.Controls;assembly=WalletWasabi.Gui"
              xmlns:cont="clr-namespace:AvalonStudio.Controls;assembly=AvalonStudio.Shell.Extensibility"
-             x:Class="WalletWasabi.Gui.Tabs.WalletManager.WalletManagerView">
+             x:Class="WalletWasabi.Gui.Tabs.WelcomeTab.WelcomeTabView">
   <Grid Classes="content">
     <DockPanel>
